# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Évite une double commande en cas de double clic",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une commande PENDING et une session de paiement",
        "tags": [
          "Commandes"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commandes"
        ]
      }
    },
    "/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Statut d'une commande (interrogé par l'app après le retour de paiement)",
        "tags": [
          "Commandes"
        ]
      }
    },
...
    "/v1/admin/orders": {
      "get": {
        "operationId": "AdminController_listOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/orders/export.csv": {
      "get": {
        "operationId": "AdminController_exportOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export CSV (sans donnée bancaire)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/orders/{id}": {
      "get": {
        "operationId": "AdminController_getOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/orders/{id}/refund": {
      "post": {
        "operationId": "AdminController_refund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users/{id}/status": {
      "patch": {
        "operationId": "AdminController_setUserStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users/{id}/export": {
      "get": {
        "operationId": "AdminController_exportUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export RGPD des données d'un utilisateur (demande transmise au support)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users/{id}/anonymize": {
      "post": {
        "operationId": "AdminController_anonymize",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Effacement RGPD : anonymise le compte, conserve les commandes",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/stats": {
      "get": {
        "operationId": "AdminController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Indicateurs : CA, commandes, panier moyen, ventes par semaine, top e-books",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/settings": {
      "get": {
        "operationId": "AdminController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      },
      "put": {
        "operationId": "AdminController_putSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/audit-logs": {
      "get": {
        "operationId": "AdminController_auditLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    }
  },
  "info": {
    "title": "API Folio",
    "description": "Vente et lecture d'e-books — application mobile et administration. Toutes les routes sont préfixées par /v1.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "ebookIds": {
            "description": "Par défaut : le contenu du panier",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "platform": {
            "type": "string",
            "enum": [
              "ios",
              "android",
              "web"
            ]
          },
          "returnUrl": {
            "type": "string",
            "description": "Adresse de retour de l'app après paiement (préfixes autorisés : CHECKOUT_RETURN_PREFIXES)"
          }
        }
      },
...
      "UserStatusDto": {
        "type": "object",
        "properties": {}
      },
      "SettingsDto": {
        "type": "object",
        "properties": {
          "appName": {
            "type": "string"
          },
          "checkoutMode": {
            "type": "object"
          },
          "webCheckoutUrl": {
            "type": "object",
            "nullable": true
          },
          "maxDevices": {
            "type": "object",
            "nullable": true
          },
          "maxDownloadsPerBook": {
            "type": "object",
            "nullable": true
          },
          "supportEmail": {
            "type": "object",
            "nullable": true
          },
          "termsUrl": {
            "type": "object",
            "nullable": true
          },
          "salesTermsUrl": {
            "type": "object",
            "nullable": true
          },
          "privacyUrl": {
            "type": "object",
            "nullable": true
          },
          "legalNoticeUrl": {
            "type": "object",
            "nullable": true
          }
        }
      }
    }
  }
}